refactor, wip: issue with navigating back to graph
