--- public/test-data.json
@@ -1,85 +1,0 @@
-{
-    "nodes": [
-      {
-        "data": {
-          "id": "foo.2000",
-          "label": "foo 2000"
-        }
-      },
-      {
-        "data": {
-          "id": "bar.2000",
-          "label": "bar 2000"
-        }
-      },
-      {
-        "data": {
-          "id": "baz.2000",
-          "label": "baz 2000"
-        }
-      },
-      {
-        "data": {
-          "id": "qux.2000",
-          "label": "qux 2000"
-        }
-      },
-      {
-        "data": {
-          "id": "pub.2000",
-          "label": "pub 2000"
-        }
-      },
-      {
-        "data": {
-          "id": "bool.2000",
-          "label": "bool 2000"
-        }
-      }
-    ],
-    "edges": [
-      {
-        "data": {
-          "source": "foo.2000",
-          "target": "bar.2000",
-          "id": "5e049c34-3dc7-4f7f-8209-b94dc5caff77"
-        }
-      },
-      {
-        "data": {
-          "source": "bar.2000",
-          "target": "baz.2000",
-          "id": "a0bf774c-5d2e-4d8d-bfd3-c3a8287ffdb3"
-        }
-      },
-      {
-        "data": {
-          "source": "bar.2000",
-          "target": "pub.2000",
-          "id": "d6fffeaf-0e51-404f-80a5-1873c5e8cd29"
-        }
-      },
-      {
-        "data": {
-          "source": "qux.2000",
-          "target": "pub.2000",
-          "id": "ff747b46-b187-4c8b-9b3f-66565bfc7dec"
-        }
-      },
-      {
-        "data": {
-          "source": "baz.2000",
-          "target": "bool.2000",
-          "id": "5be7b924-9b58-4d06-a564-5b30c179e47d"
-        }
-      },
-      {
-        "data": {
-          "source": "pub.2000",
-          "target": "bool.2000",
-          "id": "ebd4a4e5-e2db-4bed-ae33-b9003805904c"
-        }
-      }
-    ]
-  }
-
