+++ src/router/paths.js
@@ -21,15 +21,15 @@ export default [
     },
     alias: ['/', '/user/:id']
   },
-  // {
-  //   path: '/graph',
-  //   // Relative to /src/views
-  //   view: 'Graph',
-  //   meta: {
-  //     'layout': 'default'
-  //   },
-  //   alias: ['/', '/user/:id']
-  // },
+  {
+    path: '/graph',
+    // Relative to /src/views
+    view: 'Graph',
+    meta: {
+      'layout': 'default'
+    },
+    alias: ['/', '/user/:id']
+  },
   {
     path: '/dagre',
     // Relative to /src/views
