+++ src/styles/cytoscape/cytoscape-custom.css
@@ -52,7 +52,7 @@
   left: 1px;
 }
 
-.reset-button {
+.dagre-button {
   font-family: 'Avenir', Helvetica, Arial, sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
